修复bug1

+++ src/main/java/demo/controller/DynamicController.java
@@ -1,6 +1,8 @@
 package demo.controller;
 
 
+import demo.exception.BusinessException;
+import demo.exception.ErrorCode;
 import demo.service.DynamicService;
 import demo.vo.DynamicVO;
 import demo.vo.Result;
@@ -8,6 +10,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -26,7 +29,9 @@ public class DynamicController {
     @PostMapping("/createDynamic")
     @ApiOperation("创建动态")
     public Result createDynamic(@RequestBody DynamicVO dynamic, HttpServletRequest request) {
-        String userId = request.getSession().getAttribute("user").toString();
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
         dynamicService.createDynamic(dynamic, userId);
         return Result.OK().build();
     }
@@ -34,21 +39,27 @@ public class DynamicController {
     @GetMapping("/getAllDynamic")
     @ApiOperation("按时间过去所有动态")
     public Result getAllDynamic(HttpServletRequest request) {
-        String userId = request.getSession().getAttribute("user").toString();
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
         return Result.OK().data(dynamicService.selectAllDynamic(userId)).build();
     }
 
     @GetMapping("/getUserDynamic")
     @ApiOperation("查看用户自己的动态")
     public Result getUserDynamic(HttpServletRequest request) {
-        String userId = request.getSession().getAttribute("user").toString();
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
         return Result.OK().data(dynamicService.getUserDynamic(userId)).build();
     }
 
     @PostMapping("/updateDynamic")
     @ApiOperation("用户更新自己的动态")
     public Result updateUserDynamic(HttpServletRequest request, @RequestParam("dynamicId") String dynamicId, @RequestParam("content") String content) throws Exception {
-        String userId = request.getSession().getAttribute("user").toString();
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
         dynamicService.updateUserDynamic(userId, dynamicId, content);
         return Result.OK().build();
     }
@@ -56,47 +67,58 @@ public class DynamicController {
     @PostMapping("/deleteDynamic")
     @ApiOperation("删除用户的动态")
     public Result deleteUserDynamic(HttpServletRequest request, @RequestParam("dynamicId") String dynamicId) throws Exception {
-        String userId = request.getSession().getAttribute("user").toString();
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
         dynamicService.deleteUserDynamic(userId, dynamicId);
         return Result.OK().build();
     }
 
     @GetMapping("/getDynamicByGameId")
     @ApiOperation("按游戏获取动态")
-    public Result findDynamicByGameId(@RequestParam("gameId") String gameId,HttpServletRequest request) {
-        String userId = request.getSession().getAttribute("user").toString();
-        return Result.OK().data(dynamicService.findDynamicByGameId(gameId,userId)).build();
+    public Result findDynamicByGameId(@RequestParam("gameId") String gameId, HttpServletRequest request) {
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
+        return Result.OK().data(dynamicService.findDynamicByGameId(gameId, userId)).build();
     }
 
     @PostMapping("/giveDynamicLike")
     @ApiOperation("点赞")
     public Result giveDynamicLike(HttpServletRequest request, @RequestParam("dynamicId") String dynamicId) throws Exception {
-        String userId = request.getSession().getAttribute("user").toString();
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
         dynamicService.giveDynamicLikes(userId, dynamicId);
         return Result.OK().build();
     }
 
     @GetMapping("/getDynamicById")
     @ApiOperation("按动态id拿动态")
-    public Result getDynamicById(@RequestParam("dynamicId") String dynamicId,HttpServletRequest request)
-    {
-        String userId = request.getSession().getAttribute("user").toString();
-        return Result.OK().data(dynamicService.getDynamicById(dynamicId,userId)).build();
+    public Result getDynamicById(@RequestParam("dynamicId") String dynamicId, HttpServletRequest request) {
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
+        return Result.OK().data(dynamicService.getDynamicById(dynamicId, userId)).build();
     }
 
     @PostMapping("/favouritesDynamic")
     @ApiOperation("收藏动态")
-    public Result favouritesDynamic(@RequestParam("dynamicId") String dynamicId,HttpServletRequest request) throws Exception {
-        String userId = request.getSession().getAttribute("user").toString();
-        dynamicService.favoritesDynamic(userId,dynamicId);
+    public Result favouritesDynamic(@RequestParam("dynamicId") String dynamicId, HttpServletRequest request) throws Exception {
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
+        dynamicService.favoritesDynamic(userId, dynamicId);
         return Result.OK().build();
     }
 
     @PostMapping("/cancelFavouritesDynamic")
     @ApiOperation("取消收藏")
-    public Result cancelFavouritesDynamic(@RequestParam("dynamicId") String dynamicId,HttpServletRequest request) throws Exception {
-        String userId = request.getSession().getAttribute("user").toString();
-        dynamicService.deleteFavouritesDynamic(userId,dynamicId);
+    public Result cancelFavouritesDynamic(@RequestParam("dynamicId") String dynamicId, HttpServletRequest request) throws Exception {
+        String userId = (String) request.getSession().getAttribute("user");
+        if (StringUtils.isEmpty(userId))
+            throw new BusinessException(ErrorCode.BAD_REQUEST_COMMON, "用户未登录");
+        dynamicService.deleteFavouritesDynamic(userId, dynamicId);
         return Result.OK().build();
     }
